xong

+++ database/migrations/2020_05_13_044044_create_discounts_table.php
@@ -20,7 +20,7 @@ class CreateDiscountsTable extends Migration
             $table->string('code')->nullable();
             $table->string('name')->nullable();
 
-            $table->double('percent')->nullable();
+            $table->integer('percent')->nullable();
             
             $table->timestamp('created_at')
                     ->default(DB::raw('CURRENT_TIMESTAMP'))
